backend/gateway/requirements.txt
fastapi
uvicorn
motor
redis
python-multipart
pydantic
pydantic-settings
ffmpeg-python
